Improve product_create_group
With this new approach, only members of the new group will be able to create or edit products.

This can however be customized per Odoo database. Extra groups such as Purchase / Managers can be added to
the extended security rules.

The main goal of this approach is to fix bugs in Odoo related to the access of products.
For example, in some edge cases, when creating an inventory adjustment, a stack trace is
raised if the user has not access to the product.